refactor: rename ray_data module to ray for consistency
- Rename sumeh.engines.ray_data → sumeh.engines.ray
- Update all imports across __init__, core/utils, and ray module files
- Fix engine name from 'ray_data' to 'ray' in capabilities and validation reports
- Update README documentation to reflect new import path
- Add mypy configuration to ignore optional engine imports (pyflink, dask, ray)

BREAKING CHANGE: ray engine now imported as 'from sumeh import ray' instead of 'ray_data'
