+++ pyproject.toml
@@ -89,3 +89,11 @@ prerelease = false
 match = "^release/.+"
 prerelease = true
 prerelease_token = "rc"
+
+[tool.mypy]
+python_version = "3.10"
+ignore_missing_imports = false
+
+[[tool.mypy.overrides]]
+module = ["pyflink.*", "dask.*", "ray.*"]
+ignore_missing_imports = true
